Remove WeatherForecastController class entirely
The WeatherForecastController class in the QuestX.Controllers namespace has been completely removed. This includes the deletion of the ApiController and Route attributes, the ILogger dependency, the static Summaries array, and the Get method that returned a list of weather forecasts. This change eliminates all weather forecast-related API handling from the project.